+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - (Core) Renamed `output-dir` option to `output` for `snapshot` command.
 
+### Fixed
+- (Core) Path to docroot not being set using realpath (caused broken assets in snapshot)
+
 ## 1.0.4
 ### Added
 - (Core) Docroot setting added to configuration using the `setDocroot` method on the `MannequinConfig` object.